requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
dhooks==1.1.3
idna==2.8
multidict==4.7.3
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.7
yarl==1.4.2
